+++ components/Background/background.tsx
