+++ components/Window/CustomWindows/mySelfTaughtCurriculum.tsx
@@ -1,7 +1,7 @@
 import { selfTaughtCurriculumData } from "@/data/selfTaughtCurriculumData";
 import Link from "next/link";
 import { useState } from "react";
-import Window from "../Window/Window";
+import Window from "../Window";
 
 export default function MySelfTaughtCurriculumWindow(): JSX.Element {
